feat(client): refactor client into an interface
Refactored RestClient implementation into a new client_rest.go file.
Introduced Client interface in client.go for improved abstraction and testability. Updated command code to use interface and type assertions where needed.
